refactor!: restructured api to use video_player controller
