+++ example/pubspec.yaml
@@ -23,6 +23,7 @@ dependencies:
     path: ../
   flutter:
     sdk: flutter
+  video_player: ^2.10.0
 
 dev_dependencies:
   # The "flutter_lints" package below contains a set of recommended lints to
