+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: cached_video_player_plus
-description: "Original video_player plugin with the superpower of caching embedded in Android, iOS and macOS."
+description: "Original video_player plugin with the superpower of caching, because nobody likes buffering..."
 version: 3.0.3
 homepage: https://outdatedguy.rocks
 repository: https://github.com/OutdatedGuy/cached_video_player_plus
@@ -13,11 +13,7 @@ dependencies:
     sdk: flutter
   flutter_cache_manager: ^3.4.1
   get_storage: ^2.1.1
-  html: ^0.15.6
-  video_player_android: ^2.8.7
-  video_player_avfoundation: ^2.7.2
-  video_player_platform_interface: ^6.4.0
-  video_player_web: ^2.3.5
+  video_player: ^2.10.0
 
 dev_dependencies:
   flutter_lints: ^6.0.0
